requirements.txt
wheel
baseten
pypptx
flask
Flask-Login
six
